requirements.txt
# Optional: for scripts/controller_to_networktables.py (WPILib NetworkTables publisher)
# Install: pip install -r scripts/requirements-nt.txt
pynetworktables
